+++ tests/integration/conftest.py
@@ -7,7 +7,6 @@ from tests.constants import FINETUNER_LABEL
 
 import finetuner
 import hubble
-from finetuner.client import FinetunerV1Client
 
 
 @pytest.fixture()
@@ -30,7 +29,7 @@ def get_image_data():
 
 
 @pytest.fixture()
-def test_client(mocker):
+def finetuner_mocker(mocker):
     def hubble_login_mocker():
         print('Successfully logged in to Hubble!')
 
@@ -41,6 +40,5 @@ def test_client(mocker):
     mocker.patch.object(hubble.Auth, 'get_auth_token', get_auth_token)
 
     finetuner.login()
-    client = FinetunerV1Client()
 
-    return client
+    return finetuner.ft